Fixed Popup size for Android, iOS, and Windows
