+++ src/CommunityToolkit.Maui.Core/Handlers/Popup/PopUpHandler.windows.cs
@@ -103,7 +103,7 @@ public partial class PopupHandler : ElementHandler<IPopup, Popup>
 	/// <param name="view">An instance of <see cref="IPopup"/>.</param>
 	public static void MapSize(PopupHandler handler, IPopup view)
 	{
-		handler.PlatformView.SetSize(view, handler.MauiContext);
+
 	}
 
 	/// <inheritdoc/>
@@ -119,10 +119,6 @@ public partial class PopupHandler : ElementHandler<IPopup, Popup>
 		parent.IsHitTestVisible = true;
 		platformView.IsOpen = false;
 		platformView.Closed -= OnClosed;
-		if (MauiContext is not null)
-		{
-			MauiContext.GetPlatformWindow().SizeChanged -= OnSizeChanged;
-		}
 	}
 
 	/// <inheritdoc/>
@@ -137,10 +133,7 @@ public partial class PopupHandler : ElementHandler<IPopup, Popup>
 	{
 		platformView.Closed += OnClosed;
 		platformView.ConfigureControl(VirtualView, MauiContext);
-		if (MauiContext is not null)
-		{
-			MauiContext.GetPlatformWindow().SizeChanged += OnSizeChanged;
-		}
+
 		base.ConnectHandler(platformView);
 	}
 
@@ -158,15 +151,6 @@ public partial class PopupHandler : ElementHandler<IPopup, Popup>
 		if (!PlatformView.IsOpen && VirtualView.CanBeDismissedByTappingOutsideOfPopup)
 		{
 			VirtualView.Handler?.Invoke(nameof(IPopup.OnDismissedByTappingOutsideOfPopup));
-		}
-	}
-
-	void OnSizeChanged(object? sender, WindowSizeChangedEventArgs e)
-	{
-		if (VirtualView is not null)
-		{
-			PopupExtensions.SetSize(PlatformView, VirtualView, MauiContext);
-			PopupExtensions.SetLayout(PlatformView, VirtualView, MauiContext);
 		}
 	}
 }
